generalize locator
